Fix #3177 Moved geostrore api in persistence api (#3124)
